Update PostsController:
- Add Post creation
- Add access control for updatePost and removePost -- only owner can manage own posts

+++ backend/src/api/posts/posts.controller.ts
@@ -1,16 +1,35 @@
-import { Body, Controller, Delete, Get, HttpCode, Param, Patch } from '@nestjs/common';
+import {
+  Body,
+  Controller,
+  Delete,
+  ForbiddenException,
+  Get,
+  HttpCode,
+  Param,
+  Patch,
+  Post as HttpPost,
+} from '@nestjs/common';
 import { ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger';
 
+import { PqRequiresAuth } from '../shared/decorators/require-auth.decorator';
+import { PqUser } from '../shared/decorators/user.decorator';
 import { PostsService } from './posts.service';
+import { Profile } from '../profiles/entities/profile.entity';
 import { Post } from './entities/post.entity';
-import { UpdatePostDto } from './dto';
+import { CreatePostDto, UpdatePostDto } from './dto';
 
 @ApiTags('Posts')
 @Controller('/api/v1/posts')
 export class PostsController {
   constructor(private readonly postsService: PostsService) {}
 
-  // TODO
+  @HttpPost()
+  @PqRequiresAuth()
+  @ApiOperation({ summary: 'Create a new post' })
+  @ApiResponse({ status: 201, type: Post })
+  createPost(@Body() dto: CreatePostDto, @PqUser('profile') profile: Profile): Promise<Post> {
+    return this.postsService.create(profile, dto);
+  }
 
   @Get(':postId')
   @ApiOperation({ summary: 'Get a post by id' })
@@ -23,9 +42,19 @@ export class PostsController {
   @Patch(':postId')
   @ApiOperation({ summary: 'Update a post' })
   @ApiResponse({ status: 200, type: Post })
+  @ApiResponse({ status: 403, description: 'Forbidden' })
   @ApiResponse({ status: 404, description: 'Not Found' })
-  async updatePost(@Param('postId') id: string, @Body() dto: UpdatePostDto): Promise<Post> {
+  async updatePost(
+    @Param('postId') id: string,
+    @Body() dto: UpdatePostDto,
+    @PqUser('profile') profile: Profile,
+  ): Promise<Post> {
     const post = await this.getPostById(id);
+
+    // User can update only own posts
+    if (post.profile.id != profile.id) {
+      throw new ForbiddenException();
+    }
 
     return this.postsService.update(post, dto);
   }
@@ -33,10 +62,19 @@ export class PostsController {
   @Delete(':postId')
   @ApiOperation({ summary: 'Delete a post' })
   @ApiResponse({ status: 204, description: 'Empty Response' })
+  @ApiResponse({ status: 403, description: 'Forbidden' })
   @ApiResponse({ status: 404, description: 'Not Found' })
   @HttpCode(204)
-  async removePost(@Param('postId') id: string): Promise<void> {
+  async removePost(
+    @Param('postId') id: string,
+    @PqUser('profile') profile: Profile,
+  ): Promise<void> {
     const post = await this.getPostById(id);
+
+    // User can remove only own posts
+    if (post.profile.id != profile.id) {
+      throw new ForbiddenException();
+    }
 
     return this.postsService.remove(post);
   }
